fix(admin): 메시지 관련 컴포넌트 구조 개선
