+++ apps/admin/src/services/message/mutations.ts
@@ -1,6 +1,5 @@
 import { useApiError } from '@/hooks';
 import { useMutation } from '@tanstack/react-query';
-import { toast } from 'react-toastify';
 import { sendMessageByStatus, sendMessageByType, sendMessageToAll } from './api';
 import type { SendMessageByStatusRequest, SendMessageByTypeRequest, SendMessageToAllRequest } from './api';
 
@@ -9,11 +8,6 @@ export const useSendMessageByStatusMutation = () => {
 
   const { mutate: sendMessageByStatusMutate, ...restMutation } = useMutation({
     mutationFn: (params: SendMessageByStatusRequest) => sendMessageByStatus(params),
-    onSuccess: () => {
-      toast('메시지가 전송되었습니다.', {
-        type: 'success',
-      });
-    },
     onError: handleError,
   });
 
@@ -25,11 +19,6 @@ export const useSendMessageByTypeMutation = () => {
 
   const { mutate: sendMessageByTypeMutate, ...restMutation } = useMutation({
     mutationFn: (params: SendMessageByTypeRequest) => sendMessageByType(params),
-    onSuccess: () => {
-      toast('메시지가 전송되었습니다.', {
-        type: 'success',
-      });
-    },
     onError: handleError,
   });
 
@@ -41,11 +30,6 @@ export const useSendMessageToAllMutation = () => {
 
   const { mutate: sendMessageToAllMutate, ...restMutation } = useMutation({
     mutationFn: (params: SendMessageToAllRequest) => sendMessageToAll(params),
-    onSuccess: () => {
-      toast('메시지가 전송되었습니다.', {
-        type: 'success',
-      });
-    },
     onError: handleError,
   });
